Refactor out interface
Create a crate with types and host interface that doesn't depend on
other crates to prevent circular dependencies as we thread the host
interface through the core
