+++ ffi/vmlib/Cargo.toml
@@ -11,6 +11,7 @@ edition.workspace = true
 crate-type = ["cdylib", "rlib"]
 
 [dependencies]
+athena-interface = { path = "../../interface" }
 athena-runner = { path = "../../runner" }
 athcon-sys = { path = "../athcon/bindings/rust/athcon-sys" }
 athcon-vm = { path = "../athcon/bindings/rust/athcon-vm" }
