+++ interface/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "athena-interface"
+version.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+license.workspace = true
+edition.workspace = true
+
+[dependencies]
