+++ interface/src/lib.rs
@@ -1,0 +1,150 @@
+//! # Athena Interface
+//!
+//! A library with no external dependencies that includes core types and traits.
+
+use std::fmt;
+
+pub type Address = [u8; 24];
+pub type Balance = u64;
+pub type Bytes32 = [u8; 32];
+pub type Bytes = [u8];
+
+#[derive(Debug, PartialEq)]
+pub enum StorageStatus {
+  StorageAssigned,
+  StorageAdded,
+  StorageDeleted,
+  StorageModified,
+  StorageDeletedAdded,
+  StorageModifiedDeleted,
+  StorageDeletedRestored,
+  StorageAddedDeleted,
+  StorageModifiedRestored,
+}
+
+impl fmt::Display for StorageStatus {
+  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    match self {
+      StorageStatus::StorageAssigned => write!(f, "The storage item is assigned without affecting the cost structure."),
+      StorageStatus::StorageAdded => write!(f, "A new storage item is added by changing the current clean zero to a nonzero value."),
+      StorageStatus::StorageDeleted => write!(f, "A storage item is deleted by changing the current clean nonzero to the zero value."),
+      StorageStatus::StorageModified => write!(f, "A storage item is modified by changing the current clean nonzero to another nonzero value."),
+      StorageStatus::StorageDeletedAdded => write!(f, "A storage item is added by changing the current dirty zero to a nonzero value other than the original."),
+      StorageStatus::StorageModifiedDeleted => write!(f, "A storage item is deleted by changing the current dirty nonzero to the zero value and the original value is not zero."),
+      StorageStatus::StorageDeletedRestored => write!(f, "A storage item is added by changing the current dirty zero to the original value."),
+      StorageStatus::StorageAddedDeleted => write!(f, "A storage item is deleted by changing the current dirty nonzero to the original zero value."),
+      StorageStatus::StorageModifiedRestored => write!(f, "A storage item is modified by changing the current dirty nonzero to the original nonzero value other than the current."),
+    }
+  }
+}
+
+#[derive(Copy, Clone)]
+pub struct TransactionContext {
+  pub gas_price: u64,
+  pub origin: Address,
+  pub block_height: i64,
+  pub block_timestamp: i64,
+  pub block_gas_limit: i64,
+  pub chain_id: Bytes32,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum MessageKind {
+  Call,
+}
+
+#[derive(Debug, Clone, PartialEq)]
+pub struct AthenaMessage {
+  pub kind: MessageKind,
+  pub depth: i32,
+  pub gas: i64,
+  pub recipient: Address,
+  pub sender: Address,
+  pub input_data: Vec<u8>,
+  pub value: Balance,
+  pub code: Vec<u8>,
+}
+
+impl AthenaMessage {
+  pub fn new(kind: MessageKind, depth: i32, gas: i64, recipient: Address, sender: Address, input_data: Vec<u8>, value: Balance, code: Vec<u8>) -> Self {
+    AthenaMessage {
+      kind,
+      depth,
+      gas,
+      recipient,
+      sender,
+      input_data,
+      value,
+      code,
+    }
+  }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum StatusCode {
+  Success,
+  Failure,
+  Revert,
+  OutOfGas,
+  UndefinedInstruction,
+  InvalidMemoryAccess,
+  CallDepthExceeded,
+  PrecompileFailure,
+  ContractValidationFailure,
+  ArgumentOutOfRange,
+  InsufficientBalance,
+  InternalError,
+  Rejected,
+  OutOfMemory,
+}
+
+impl fmt::Display for StatusCode {
+  fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+    match self {
+      StatusCode::Success => write!(f, "Execution finished with success."),
+      StatusCode::Failure => write!(f, "Generic execution failure."),
+      StatusCode::Revert => write!(f, "Execution terminated with REVERT opcode."),
+      StatusCode::OutOfGas => write!(f, "The execution has run out of gas."),
+      StatusCode::UndefinedInstruction => write!(f, "An undefined instruction has been encountered."),
+      StatusCode::InvalidMemoryAccess => write!(f, "Tried to read outside memory bounds."),
+      StatusCode::CallDepthExceeded => write!(f, "Call depth has exceeded the limit."),
+      StatusCode::PrecompileFailure => write!(f, "A call to a precompiled or system contract has ended with a failure."),
+      StatusCode::ContractValidationFailure => write!(f, "Contract validation has failed."),
+      StatusCode::ArgumentOutOfRange => write!(f, "An argument to a state accessing method has a value outside of the accepted range."),
+      StatusCode::InsufficientBalance => write!(f, "The caller does not have enough funds for value transfer."),
+      StatusCode::InternalError => write!(f, "Athena implementation generic internal error."),
+      StatusCode::Rejected => write!(f, "The execution of the given code and/or message has been rejected by the Athena implementation."),
+      StatusCode::OutOfMemory => write!(f, "The VM failed to allocate the amount of memory needed for execution."),
+    }
+  }
+}
+
+#[derive(Debug)]
+pub struct ExecutionResult {
+    pub status_code: StatusCode,
+    pub gas_left: i64,
+    pub output: Option<Vec<u8>>,
+    pub create_address: Option<Address>,
+}
+
+impl ExecutionResult {
+  pub fn new(status_code: StatusCode, gas_left: i64, output: Option<Vec<u8>>, create_address: Option<Address>) -> Self {
+    ExecutionResult {
+      status_code,
+      gas_left,
+      output,
+      create_address,
+    }
+  }
+}
+
+pub trait HostInterface {
+  fn account_exists(&self, addr: &Address) -> bool;
+  fn get_storage(&self, addr: &Address, key: &Bytes32) -> Bytes32;
+  fn set_storage(&mut self, addr: &Address, key: &Bytes32, value: &Bytes32) -> StorageStatus;
+  fn get_balance(&self, addr: &Address) -> Balance;
+  // this should take an opaque context object (txid? something else?)
+  fn get_tx_context(&self) -> TransactionContext;
+  fn get_block_hash(&self, number: i64) -> Bytes32;
+  fn call(&mut self, msg: AthenaMessage) -> ExecutionResult;
+}
