+++ runner/Cargo.toml
@@ -11,4 +11,5 @@ edition.workspace = true
 athcon-client = { path = "../ffi/athcon/bindings/rust/athcon-client" }
 athcon-sys = { path = "../ffi/athcon/bindings/rust/athcon-sys" }
 athcon-vm = { path = "../ffi/athcon/bindings/rust/athcon-vm" }
+athena-interface = { path = "../interface" }
 athena-sdk = { path = "../sdk" }
