+++ runner/src/lib.rs
@@ -2,17 +2,7 @@ pub mod host;
 pub mod vm;
 
 pub use host::{
-  Address,
-  AthenaMessage,
-  Balance,
-  Bytes32,
   Bytes32AsU64,
   ExecutionContext,
-  ExecutionResult,
-  HostInterface,
-  MessageKind,
-  StatusCode,
-  StorageStatus,
-  TransactionContext,
 };
 pub use vm::{AthenaVm, VmInterface};
