+++ sdk/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 serde = { version = "1.0.201", features = ["derive"] }
 anyhow = "1.0.83"
 athena-core = { path = "../core" }
+athena-interface = { path = "../interface" }
 bincode = "1.3.3"
 tracing = "0.1.40"
 hex = "0.4.3"
